RF: extend CV section environment; add subsection
I found it useful to a) be able to add stuff to the end of the section
title (a footnote in fact), and b) have subsections within the CV
sections.  So, I added an optional argument to the section environment
to add extra markup at the end of the section title, and added a
subsection enviroment.